[Win32] lexer - fix warnings
remaining is a length, so a size_t

ptrdiff_t shall be used for the arithmetics with pointers.

+++ common/lexer.h
@@ -49,7 +49,7 @@ enum {
 typedef struct {
 	char *filename;
 	const char *at;
-	int remaining;
+	size_t remaining;
 	int complained;
 
 	int tok_type;
